- FSIOperator & FSISolver reordered in sections: Constructors, Methods, Get Functions, Set Functions, etc.. - Removed some function duplications; - Added "const" in input parameters (where it was possible); - Removed "ifndef TWODIM" inside functions leaving only the external ones that are enough to hide everything; - Removed FSIOperator leaderPrint functions: we now use the ones defined in the Displayer template class - Other minor and cosmetics changes..
NOTE: It is possible that if you have some uncommitted changes inside FSIOperator & FSISolver classes, your version may have some conflicts with this new version due to the fact that the order of the functions in the headers has been changed. However, function content is almost the same, so a manual merge should be easy.