Moving the registration of preconditioners and FSIoperator methods into their specific classes.
On my machine this is working fine with static compilation of the library (which is the critical one for cluster simulation). It has been tested both in debug and opt2 modes.

For now this affects only test_fsi. If it works well on other machines, this will be expanded to all the other tests inside the testsuite.