Doxygen files and configuration updated.
Now we have a better version, however I was not able to fix a couple of things into the configurations. For instance, I don't know why there is still a double copy of the index menu in each main page of the documentation (that are "life/life*/").

However, for now it is not a big problem.