Exporters updated:
1) Reordering exporter.hpp and .cpp
2) Moving some methods and members from exporter to ensight as they are used only there
3) Added a new method: importFromTime(t). Basically this is intended for replacing the existing import(t) and import(Tstart,t) method.
The idea is the same, but the implementation is completely different. In particular, given a time and an xmf file (for hdf5, but it can be implemented also for ensight), it read the file, saving in a vector of pairs all the time steps and the iteration numbers. Then, using this information, it match the time closest to the given one (we can expect to have a difference due to roundoff), and it loads that solution. It returns the iteration number corresponding to that time, which can be passed to the new exporter (using setStartIndex), automatizing most of the operations.