Multiscale solver updated:
1) moved the part of the creation of the problem folder in the main of the test, as requested by Gilles ;)
2) implemented the restart of simulations using the new features of HDF5exporter.
In the main data file there is now a new section: Restart.
To perform a restart we should specify only few information:
a) Restart = true
b) RestartFromStepNumber = the step from which we perform the restart (corresponding to the numeration of the files)
Then we only specify the new time parameters, where initial time is the time of the previous simulation from which we want to restart, and deltaT can be different from the previous simulation.