Dear LifeVers,
I have made some changes in the core of lifefem that I am going to commit. The goal was to clarify some parts of the code and make them more flexible. I will make a commit file-by-file so that I can comment the changes.

Please, wait that I have commited everything before updating!



I start with the quadRule files:

* Added full documentation
* Conforming classes to the Guidelines
* Deleting the SetOfQuadRule class

Note that some accesses to the variables have changed for the QuadRule class (all the members are private, use the getters).