Mantaining the obsolete test_bdf.
Moving everything from MSRMatr and boost::vector to Epetra, and from aztec to Trilinos.

The new bdfVS is tested for a problem with analytical solution quadratic in space and time. The discretization error is (as expected) of the same order of the linear system tolerance in the case of P2 fem + BDF2.

I had some issue with the recomputing of the preconditioner. I'll explain my problem in lifev-dev.