understood and fixed the reason of a segmentation fault while trying to recompute the ML_Preconditioner between an iteration and the other.
The bug seems to be in the Trilinos_10.2 library (I have not tested with previous version): the destructor of a ML_Preconditioner object tryes to access an Epetra matrix pointer which can be dandling...

The walkaround to this Trilinos bug is easy to implement in LifeV::MLPreconditioner.
M_Oper is indeed a shared_ptr which owns the matrix that ML tryes to access in the destructor.
Therefore simply calling first  M_Prec.reset() and then M_Oper.reset(), we ensure that the not safe pointer in Trilinos::ML_Preconditioner is indeed always valid.

Did anybody else sperimented similar problems with ML?