Walkaround to fix a bug in Trilinos::MultiLevelPreconditioner (version 10.2, and maybe previous versions).
Trilinos::MultiLevelPreconditioner contains a not safe pointer to an Epetra_CsrMatrix.

The matrix in question is actually co-owed by the shared_ptr M_Oper in LifeV::MLPreconditioner.

If we take care to always destroy M_Prec (which is a shared_ptr to Trilinos::MultiLevelPreconditioner) first and then to destroy M_Oper we are sure that area of memory which is not safetly accessed by Trilinos is actually valid.

Therefore I switched the order of the two resets in MLPrecondtioner::precReset(), and i called first M_Prec.reset(); in MLPreconditioner::buildPreconditioner(operator &).