Changed the internal structure of partitionMesh class.
This is the new version of the partitionMesh class. There are no changes to the interface, just a modification of the internal structure and added doxygen comments.
The mesh partitioning used to be done in the constructor. In the new version, the constructor only calls a private method, partitionMesh::execute, which handles mesh partitioning. Each phase of the process has been placed in a separate private method, so it is clearer to understand what happens. The partitionMesh::execute calls each of these methods in sequence, it doesn't do anything else.
In the future, it could be better to remove the call to the execute method from the constructor and make execute a separate public method.