New data container for FSI. This class contains the all the data for FSI problem, plus a shared_ptr to the data of the fluid and of the solid.
IMPORTANT NOTE: The greatest part of the data of FSISolver, FSIOperator (and derived classes) have been moved here. There are still some data that cannot be moved now. For them, a new method setDataFile has been added in place of the old setDataFromGetpot. This method should be removed in the future, because it is just a legacy due to some interfaces which now are not ready for the conversion.