Changed the way in which generalizedAitken check if the computed Omega is acceptable or not. Before the ammissible range was:
abs(defaultOmega)/1024 < abs(Omega) < abs(defaultOmega)*1024

and if Omega was out of the two bounds, it was re-initialized to omegaDefault (which was always positive)

Now we provide (through a new method) two limits: OmegaMin and OmegaMax.
The new ammissible range is:

abs(OmegaMin) < abs(Omega) < abs(OmegaMax)

and if Omega is outside this range, it assume the maximum/minimum ammissible value keeping the sing:
e.g. OmegaMin=0.2, OmegaMax=0.5, OmegaDefault=0.35, Omega=-0.9 ==> New Omega = -0.5 (instead of 0.35).