In order to go in the direction of the 1D and 2D LifeV, changes have been done in the way the basis functions are accessed: instead of (x,y,z) or (x,y) or (x) depending on the case, it uses now a GeoVector (that is currently a typedef for boost::numeric::ublas::vector, see in lifearray/tab.hpp file).
Old style accessors using (x,y?,z?) are however still available (but might be slower because
of the convertion) => No change in the codes needed.

Precision: this change is ONLY for the access of the basis functions and some closely related
methods (in lifefem). It does not affects for example the definition of the boundary conditions.