BUGFIX in EpetraVector::add(EpetraVector & v, UInt offset)
Even if the offset is 0
you can't add vector with different lenght using operator+=, but you need to sum component by component as done for offset != 0.

Infact Epetra_MultiVector::Update (which is the function called by LifeV::EpetraMatrix::operator+=) will generate an error (-2) if the lenght of the local vectors are different and will return without summing.

Is any developer using this methods?