I noticed that some template function in bcManage.hpp will not compile anymore when using EpetraMatrix.
All this function where written to handle the b.c. in the prehistoric NavierStokesSolverPC, and assume a method zero_row(...) which EpetraMatrix does not provide...

I moved them in a new file bcHandlerObsolete.hpp and double checked that the moved functions are not use anywhere in the library.

If nobody reports any problem in their playgrounds I suggest to permanentely delete the file bcHandlerObsolete.hpp.