BUGFIX for assembling a rectangular matrix in parallel.
It turned out that the correct way to deal with rectangular FEM
matrices in trilinos is to specify only the row map in the contructor.
Then for shaping the matrix correctly we should use the function
Epetra_FeCsrMatrix::GlobalAssemble(OperatorDomainMap,
OperatorRangeMap);

Here there is the list of the modification I will commit:

1) The constructor which takes in input two maps has been removed.
2) Any occurrence of M_colMap has been removed.
3) M_rowMap is called simply M_map;
4) two new shared_ptr<LifeV::EpetraMap> has been added. M_domainMap, M_rangeMap.
These two pointer will be NULL pointer until GlobalAssemble() is called.
5) The GlobalAssemble(...) for rectangular matrices now takes two shared_ptr
5) Added more documentiation and ASSERT_PRE.

Since I believe none of you was using RECTANGULAR matrices,
the interface changes should not affect your code.