fixing a compilation issue (in opt mode)
list_Type was actually not defined in ComposedPreconditioner.
The current implementation requires the creation of a temporary object of
type type_List, to be passed to createList. However, the list object is not
used. I marked it as "deprecated", so that we can remember to look for
a better implementation.