Temporary fix on the material flags.
Until Paolo Tricerri finishes his development on the structural solver
I commit a temporary fix to restore the full functionality for the
case of 1 single material. In particular:
 - A new boolean variable M_materialsFlagSet has been introduced. It
is set to false if the string "material_flag" is not provided in the
data file, while it is true in the other case.
 - If a material flag is provided, the data container works as before.
 - If no material flag is provided, the data container store the Young
modulus and the Poisson in ID = 1. The getters use the
M_materialFlagSet variable to understand if they have to provide the
quantities for the user provided material or if they should return the
values stored in ID = 1.
 
 This approach is compatible with old code (no changes have been done
in the testsuite), but it allows more flexibility (especially for the
Multiscale framework).

Additional minor changes to the class:
* The variables M_factor, M_verbose, M_order, and M_solidType were
wrongly setted due to missing "section" in the getpot string;
* Inlined some setters and getters;
* Some cleaning and renaming;