BugFix / Cleaning abs VS fabs
This commit aims to fix a subtle bug in the call to the function abs().

According to the include header (cmath or cstdlib) the function
int abs(int) has an overloaded function double abs(double) (in cmath) or not.

As a result, if the wrong header is included (i.e. if cstdlib is included instead of cmath)
then a down cast from double to int may happen when abs(x) is called,
introducing a bug.

A clean solution, implemeted in this commit is to always use:
int std::abs(int) when working with integers
double std::fabs(double) when working with doubles.