Extract a 2d mesh from the boundary of a 3d mesh
The mesh extraction procedure still have some bugs.
In particular, the number of boundary edges is not computed correctly.

The 2dmesh build with this code is not perfectly compatible
with the meshPartioner due to the bug above.

In Debug mode an ASSERT is not satisfied, while in Opt mode a segmentation fault
happen if we try to divide the 2d mesh in too many subdomains.