using debugStream() rather than Debug()
"Debug" used to be the name of LifeV's debug stream (an object
of type DebugStream. When the macro NDEBUG_OLD was defined,
the symbol "Debug" was converted to "NDebug" to suppress
the I/O operation. This substitution is dangerous, as it is
not unlikely that other software projects (e.g. Trilinos)
use the symbol "Debug".

To avoid clash with other code #included in the compilation
unit, the objects Debug() and NDebug() have been renamed into
debugStream() and noDebugStream()