New in ETA: Building matrices with a precomputed graph
This commits add support for precomputing matrix graphs
for building FE system matrices in a closed optimized form

Two new files: lifev/eta/expression/{BuildGraph,GraphElement}.hpp
handle precomputing the matrix graph

Changes in other files of ETA accomodate the new strategy and
selection of the open/closed path in an efficient manner.

Additions to the MatrixEpetra and MatrixEpetraStructuredView are
related to the usage of sumIntoGlobalValues, where appropriate,
without having an "if" clause in the FE loop.