Initial working version of multi-threaded assembly for ETA
This commit introduces a way to perform multi-threaded assembly
using the ETA strategy.

It provides an overload  of the integrate(...) function which
takes an additional argument, the number of threads used for
assembly. The multi-threading mechanism is based on OpenMP,
so for it to work, LifeV needs to be compiled with the appropriate
compiler switch.

Due to the fact that certain methods that make up the FE loop in
ETA had to be modified, in order to make the reentrant, the
single-threaded path has been left unchanged. Reentrant methods
are needed for the multi-threaded approach, but are slower than
non-reentrant methods, when only 1 threads is used.