Cleanup of multi-threading in ETA
Adding _OPENMP include guards to compile without OpenMP too.

In IntegrateMatrixElement, removing the dual path addToClosed,
addToClosedMT, since it's not needed anymore. The reentrant
method to integrate and element is not slower than the
non-reentrant one, which is removed.

Also removed are some data members of the IntegrateMatrixElement
class, being replaced with variables local to methods.