Adding constructor to MatrixEpetra and updating mt_assembly test
Adding a constructor that accepts an Epetra_FECrsGraph object as
parameter and has support for the overlap flag. This is useful
in conjunction with my patches to Epetra_FECrsMatrix that make
that matrix avare of thread-safety and overlap.

Update the mt_assembly test to support and use overlapping
mesh parts.

Conflicts:
	lifev/core/array/MatrixEpetra.hpp