Using Epetra_MpiComm objects in PartitionIO and HDF5IO
Modified the interface of the PartitionIO and HDF5IO classes
to use Epetra_MpiComm objects explicitly. These classes are
only enabled if LifeV is configured with MPI and HDF5 and they
are only used with MPI.

If the user of the classes need to create an Epetra_Comm object
in his code, he should explicitly cast it to an Epetra_MpiComm
when using the PartitionIO class.