Bug fix in RegionMesh::updateElementRidges
When we use the method reorderAccordingToFlag in MeshEntityContainer
we then need to adjust the localId of the ridges and also the map
M_ElemToRidge. The method resetId of MeshEntityContainer resets the localId
of the entities so that it matches the position in the container, and
returns a map newToOld (for the localId's). However, resetId is already
called inside reorderAccordingToFlag, so a second call to resetId after
reorderAccordingToFlag has been called would simply return an identity
map. Lucky enough, reorderAccordingToFlag returns the newToOld map, so we
can simply use that.